requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-debug-toolbar==4.4.6
django-parler==2.3
django-rosetta==0.10.0
idna==3.7
numpy==2.0.1
pillow==10.4.0
pip==24.2
polib==1.2.0
requests==2.32.3
rosetta==0.3
scipy==1.14.0
setuptools==70.3.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
